[RESOLVED] [ISSUE] - Hashcode must be implemented along with Equals
- In Java, Object.hashCode() behaves similarly to Object.equals(); if two objects are equal, they are guaranteed to have the same hash code.

This classes appears to define an equality condition but uses the hashCode implementation defined in AbstractEntity.class . This could result in two objects that appear to be equal but do not have the same hash code; a violation of the contract. In the more common case, this would result in a greater chance of two objects having the same hash code, which reduces the performance of collections like HashMap.